requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
alembic
PyMySQL
python-jose
passlib[bcrypt]
cachetools